Moved the functor that checks the existence of app-defined functions
